+++ firmware/hdl/payload/gtl/bx_pipeline.vhd
@@ -2,6 +2,7 @@
 -- Pipeline for +/-2 bx data.
 
 -- Version-history:
+-- HB 2021-06-10: updated for hadronic shower trigger (mus).
 -- HB 2021-05-21: first design.
 
 library ieee;
@@ -25,6 +26,11 @@ architecture rtl of bx_pipeline is
 
     type cent_tmp_array is array (0 to BX_PIPELINE_STAGES-1) of std_logic_vector(NR_CENTRALITY_BITS-1 downto 0);
     signal centrality_tmp : cent_tmp_array;
+
+    type mus_tmp_array is array (0 to BX_PIPELINE_STAGES-1) of std_logic_vector(NR_MUS_BITS-1 downto 0);
+    signal mus_tmp, mus_tmp_del : mus_tmp_array;
+
+    signal mus0, mus1, musoot0, musoot1 : mus_bit_array;
 
 begin
 
@@ -89,6 +95,25 @@ begin
             port map(
                 clk, data_tmp(i).ext_cond, bx_data.ext_cond(i)
             );
+
+        mus_tmp(i)(0) <= data_tmp(i).mu(MUON_OBJ_MUS0)(MUS_BIT);
+        mus_tmp(i)(1) <= data_tmp(i).mu(MUON_OBJ_MUS1)(MUS_BIT);
+        mus_tmp(i)(2) <= data_tmp(i).mu(MUON_OBJ_MUSOOT0)(MUS_BIT);
+        mus_tmp(i)(3) <= data_tmp(i).mu(MUON_OBJ_MUSOOT1)(MUS_BIT);
+
+        mus_pipe_i: entity work.delay_pipeline
+            generic map(
+                DATA_WIDTH => NR_MUS_BITS,
+                STAGES => MUS_STAGES
+            )
+            port map(
+                clk, mus_tmp(i), mus_tmp_del(i)
+            );
+
+        bx_data.mus0(i) <= mus_tmp_del(i)(0);
+        bx_data.mus1(i) <= mus_tmp_del(i)(1);
+        bx_data.musoot0(i) <= mus_tmp_del(i)(2);
+        bx_data.musoot1(i) <= mus_tmp_del(i)(3);
 
     end generate bx_l;