+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.15.1] - 2021-06-10
+### Comment
+
+- mp7_ugt firmware release v1.15.1
+- added logic for hadronic shower triggers (muon)
+
+### Changed
+- source files:
+  - ../payload/gtl/bx_pipeline.vhd
+  - ../packages/gtl_pkg.vhd
+  - ../packages/gt_mp7_core_pkg.vhd
+
 ## [v1.15.0] - 2021-06-09
 ### Comment